requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-filter==23.4
psycopg[binary]==3.1.13
whitenoise==6.6.0
gevent==23.9.1
django-twilio==0.14.3.2
channels[daphne]==4.0.0
channels-redis==4.2.0
pytest-django==4.7.0
pytest-asyncio==0.23.2
djangorestframework-simplejwt==5.3.1
django-rest-swagger==2.2.0
drf_yasg==1.21.7
gunicorn==22.0.0
uvicorn[standard]==0.30.1
django-cors-headers==4.3.1
dj-database-url==2.2.0
django-channels-jwt==0.0.3
